Process input

+++ src/main/java/com/itemis/bst/model/ProcessResponse.java
@@ -3,12 +3,13 @@ package com.itemis.bst.model;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @AllArgsConstructor
 @Getter
 public class ProcessResponse {
     private List<Item> itemList;
-    private double totalTaxes;
-    private double total;
+    private BigDecimal totalTaxes;
+    private BigDecimal total;
 }
